Move device authentication and connection lifecycle behind DeviceLink
DeviceSocket held the whole authentication story: certificate lookup, HMAC
header decoding, signature verification, JITP device creation, the duplicate
identifier notification, and the mapping from a Device onto a DeviceInfo. None
of that can run next to the socket once the socket lives in an application
without a database.

It now lives in NervesHub.DeviceLink.Authentication, reached as
DeviceLink.authenticate/1 taking {:ssl_cert, der} or {:shared_secret, headers}.
The HMAC check has to move whole rather than in pieces -- verifying the
signature needs the shared secret itself -- so a caller with no database hands
the headers over and gets back a DeviceInfo or :invalid_auth. The failure detail
is recorded as telemetry where it is known instead of travelling back to a
caller that cannot do anything with it.

The connection lifecycle follows: DeviceLink.connect/1 records the connection
and stamps the reference onto the device info, with heartbeat/1 and disconnect/2
keyed by that reference.

shared_secrets_enabled?/0 moves too, since authentication depends on it;
Live.Product.Settings now asks DeviceLink. The config key stays as
NervesHubWeb.DeviceSocket -- it is a deployment contract that
DEVICE_SHARED_SECRETS_ENABLED maps onto, and renaming it risks silently
disabling shared secrets.

What stays on the socket is what needs the connection: evicting a duplicate,
the serializer's device_id, heartbeat throttling, and connect/disconnect
telemetry. device_socket.ex, device_channel.ex, extensions_channel.ex and
console_channel.ex now reference no context module other than DeviceLink.
