+++ lib/nerves_hub_web/channels/device_socket.ex
@@ -2,22 +2,12 @@ defmodule NervesHubWeb.DeviceSocket do
   use Phoenix.Socket
   use OpenTelemetryDecorator
 
-  alias NervesHub.DeviceLink.DeviceInfo
-  alias NervesHub.Devices
-  alias NervesHub.Devices.Certificates
-  alias NervesHub.Devices.Connections
-  alias NervesHub.Devices.DeviceConnection
-  alias NervesHub.ProductNotifications
-  alias NervesHub.Products
+  alias NervesHub.DeviceLink
   alias Phoenix.Socket.Transport
-  alias Plug.Crypto
 
   channel("console", NervesHubWeb.ConsoleChannel)
   channel("device:*", NervesHubWeb.DeviceChannel)
   channel("extensions", NervesHubWeb.ExtensionsChannel)
-
-  # Default 90 seconds max age for the signature
-  @default_max_hmac_age 90
 
   defoverridable init: 1, handle_in: 2, terminate: 2
 
@@ -74,7 +64,7 @@ defmodule NervesHubWeb.DeviceSocket do
   @decorate with_span("Channels.DeviceSocket.heartbeat")
   defp heartbeat(%{assigns: %{device_info: device_info}} = socket) do
     if update_heartbeat?(socket) do
-      Connections.device_heartbeat(device_info.connection_ref)
+      _ = DeviceLink.heartbeat(device_info.connection_ref)
       update_last_heartbeat(socket)
     else
       socket
@@ -99,89 +89,19 @@ defmodule NervesHubWeb.DeviceSocket do
   @impl Phoenix.Socket
   @decorate with_span("Channels.DeviceSocket.connect:cert_auth")
   def connect(_params, socket, %{peer_data: %{ssl_cert: ssl_cert}}) when not is_nil(ssl_cert) do
-    X509.Certificate.from_der!(ssl_cert)
-    |> Certificates.get_device_by_x509()
-    |> case do
-      {:ok, device} ->
-        socket_and_assigns(socket, device)
-
-      error ->
-        :telemetry.execute([:nerves_hub, :devices, :invalid_auth], %{count: 1}, %{
-          auth: :cert,
-          reason: error
-        })
-
-        {:error, :invalid_auth}
+    case DeviceLink.authenticate({:ssl_cert, ssl_cert}) do
+      {:ok, device_info} -> socket_and_assigns(socket, device_info)
+      {:error, reason} -> {:error, reason}
     end
   end
 
   # Used by Devices connecting with HMAC Shared Secrets
   @decorate with_span("Channels.DeviceSocket.connect:shared_secrets")
-  def connect(_params, socket, %{x_headers: x_headers})
-      when is_list(x_headers) and (is_list(x_headers) and x_headers != []) do
-    headers = Map.new(x_headers)
-
-    with :ok <- check_shared_secret_enabled(),
-         {:ok, key, salt, verification_opts} <- decode_from_headers(headers),
-         {:ok, auth} <- get_shared_secret_auth(key),
-         {:ok, signature} <- Map.fetch(headers, "x-nh-signature"),
-         {:ok, identifier} <- Crypto.verify(auth.secret, salt, signature, verification_opts),
-         {:ok, device} <- get_or_maybe_create_device(auth, identifier) do
-      socket_and_assigns(socket, device)
-    else
-      {:error,
-       %Ecto.Changeset{
-         changes: %{identifier: identifier, org_id: org_id, product_id: product_id},
-         errors: [
-           identifier: {_msg, [constraint: :unique, constraint_name: "devices_identifier_index"]}
-         ]
-       }} ->
-        _ =
-          ProductNotifications.create_duplicate_device_identifier_notification!(
-            product_id,
-            identifier,
-            :shared_secret
-          )
-
-        :telemetry.execute([:nerves_hub, :devices, :invalid_auth], %{count: 1}, %{
-          auth: :shared_secrets,
-          reason: :duplicate_device_identifier,
-          org_id: org_id,
-          product_id: product_id,
-          identifier: identifier
-        })
-
-        {:error, :invalid_auth}
-
-      {:error, :expired} ->
-        :telemetry.execute([:nerves_hub, :devices, :invalid_auth], %{count: 1}, %{
-          auth: :shared_secrets,
-          reason: :signature_expired,
-          shared_key: Map.get(headers, "x-nh-key", "*empty*")
-        })
-
-        {:error, :invalid_auth}
-
-      error ->
-        :telemetry.execute([:nerves_hub, :devices, :invalid_auth], %{count: 1}, %{
-          auth: :shared_secrets,
-          reason: error,
-          shared_key: Map.get(headers, "x-nh-key", "*empty*")
-        })
-
-        {:error, :invalid_auth}
+  def connect(_params, socket, %{x_headers: x_headers}) when is_list(x_headers) and x_headers != [] do
+    case DeviceLink.authenticate({:shared_secret, Map.new(x_headers)}) do
+      {:ok, device_info} -> socket_and_assigns(socket, device_info)
+      {:error, reason} -> {:error, reason}
     end
-  rescue
-    e in ArgumentError ->
-      headers = Map.new(x_headers)
-
-      :telemetry.execute([:nerves_hub, :devices, :invalid_auth], %{count: 1}, %{
-        auth: :shared_secrets,
-        reason: e,
-        shared_key: Map.get(headers, "x-nh-key", "*empty*")
-      })
-
-      {:error, :invalid_auth}
   end
 
   def connect(_params, _socket, _connect_info) do
@@ -202,79 +122,9 @@ defmodule NervesHubWeb.DeviceSocket do
     ]
   end
 
-  defp decode_from_headers(%{"x-nh-alg" => "NH1-HMAC-" <> alg} = headers) do
-    with [digest_str, iter_str, key_len_str] <- String.split(alg, "-"),
-         digest = String.to_existing_atom(String.downcase(digest_str)),
-         {iterations, ""} <- Integer.parse(iter_str),
-         {key_length, ""} <- Integer.parse(key_len_str),
-         {signed_at, ""} <- Integer.parse(headers["x-nh-time"]),
-         {:ok, key} <- Map.fetch(headers, "x-nh-key") do
-      expected_salt = """
-      NH1:device-socket:shared-secret:connect
-
-      x-nh-alg=NH1-HMAC-#{alg}
-      x-nh-key=#{key}
-      x-nh-time=#{signed_at}
-      """
-
-      opts = [
-        key_length: key_length,
-        key_iterations: iterations,
-        key_digest: digest,
-        signed_at: signed_at,
-        max_age: max_hmac_age()
-      ]
-
-      {:ok, key, expected_salt, opts}
-    end
-  end
-
-  defp decode_from_headers(_headers), do: {:error, :headers_decode_failed}
-
-  defp get_shared_secret_auth("nhp_" <> _ = key), do: Products.get_shared_secret_auth(key)
-  defp get_shared_secret_auth(key), do: Devices.get_shared_secret_auth(key)
-
-  defp get_or_maybe_create_device(%Products.SharedSecretAuth{} = auth, identifier) do
-    # TODO: Support JITP profile here to decide if enabled or what tags to use
-    Devices.get_or_create_device(auth, identifier)
-  end
-
-  defp get_or_maybe_create_device(%{device: %{identifier: identifier} = device}, identifier), do: {:ok, device}
-
-  defp get_or_maybe_create_device(_auth, _identifier), do: {:error, :bad_identifier}
-
-  defp max_hmac_age() do
-    Application.get_env(:nerves_hub, __MODULE__, [])
-    |> Keyword.get(:max_age, @default_max_hmac_age)
-  end
-
-  defp check_shared_secret_enabled() do
-    if shared_secrets_enabled?() do
-      :ok
-    else
-      {:error, :shared_secrets_not_enabled}
-    end
-  end
-
-  defp socket_and_assigns(socket, device) do
+  defp socket_and_assigns(socket, device_info) do
     # disconnect devices using the same identifier
-    _ = socket.endpoint.broadcast_from(self(), "device_socket:#{device.id}", "disconnect", %{})
-
-    if device.status == :registered do
-      Devices.set_as_provisioned!(device)
-    end
-
-    device_info = %DeviceInfo{
-      org_id: device.org_id,
-      product_id: device.product_id,
-      device_id: device.id,
-      device_identifier: device.identifier,
-      deployment_id: device.deployment_id,
-      firmware_metadata: device.firmware_metadata,
-      device_updates_enabled: device.updates_enabled,
-      device_updates_blocked_until: device.updates_blocked_until,
-      allowed_extensions: calculate_allowed_extensions(device)
-    }
+    _ = socket.endpoint.broadcast_from(self(), "device_socket:#{device_info.device_id}", "disconnect", %{})
 
     {:ok, assign(socket, :device_info, device_info)}
   end
@@ -282,11 +132,10 @@ defmodule NervesHubWeb.DeviceSocket do
   @decorate with_span("Channels.DeviceSocket.on_connect")
   defp on_connect(%{assigns: %{device_info: device_info}} = socket) do
     # Report connection and use connection id as reference
-    {:ok, %DeviceConnection{id: connection_id}} =
-      Connections.device_connecting(device_info.org_id, device_info.product_id, device_info.device_id)
+    {:ok, device_info} = DeviceLink.connect(device_info)
 
     :telemetry.execute([:nerves_hub, :devices, :connect], %{count: 1}, %{
-      ref_id: connection_id,
+      ref_id: device_info.connection_ref,
       identifier: device_info.device_identifier,
       firmware_uuid: get_in(device_info, [Access.key(:firmware_metadata), Access.key(:uuid)])
     })
@@ -300,7 +149,7 @@ defmodule NervesHubWeb.DeviceSocket do
     Process.put(:device_id, device_info.device_id)
 
     socket
-    |> assign(device_info: %{device_info | connection_ref: connection_id})
+    |> assign(:device_info, device_info)
     |> update_last_heartbeat()
   end
 
@@ -328,16 +177,9 @@ defmodule NervesHubWeb.DeviceSocket do
 
     # its possible that this is a stale connection and the device has already reconnected,
     # which means the following call might return :error, but we can ignore it
-    _ = Connections.device_disconnected(device_info.connection_ref)
+    _ = DeviceLink.disconnect(device_info.connection_ref)
 
     assign(socket, :disconnection_handled?, true)
-  end
-
-  defp calculate_allowed_extensions(device) do
-    for {extension, true} <- Map.from_struct(device.product.extensions),
-        {^extension, device_enabled?} <- Map.from_struct(device.extensions),
-        device_enabled? != false,
-        do: extension
   end
 
   defp last_seen_update_interval() do
@@ -346,11 +188,5 @@ defmodule NervesHubWeb.DeviceSocket do
     jitter = Application.get_env(:nerves_hub, :device_last_seen_update_interval_jitter_seconds)
 
     interval + Enum.random(-jitter..jitter)
-  end
-
-  def shared_secrets_enabled?() do
-    Application.get_env(:nerves_hub, __MODULE__, [])
-    |> Keyword.get(:shared_secrets, [])
-    |> Keyword.get(:enabled, false)
   end
 end
